+++ remix/app/routes/api/v1/vercel/status-data/_status-data.tsx
@@ -1,8 +1,12 @@
 import { json } from '@vercel/remix';
 
-import { getVercelDeploymentStatus } from '~/api/utils/vercel/status';
+import { getVercelDeploymentStatus, isVercelStatusEnabled } from '~/api/utils/vercel/status';
 
 export const loader = async () => {
+  if (!isVercelStatusEnabled()) {
+    throw new Response('Not found', { status: 404 });
+  }
+
   const status = await getVercelDeploymentStatus();
   return json(status);
 };